Improve monorepo task workflows
* Faster task and CI builds by excluding 'app' workspace builds.
* Faster code formatting by excluding large fixtures.
* Automatic precommit checks for format and lint errors.
* 'npm test' runs all the same tests as the CI workflow.
* Minor reorganization of workflow jobs.
