--- .github/workflows/typos.yml
@@ -1,15 +1,0 @@
-name: Spell Check 
-on:
-  pull_request:
-  push:
-    branches:
-      - "main"
-jobs:
-  typos:
-    name: Spell Check with Typos
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout Actions Repository
-        uses: actions/checkout@v4
-      - name: Check spelling
-        uses: crate-ci/typos@master
