+++ apps/esbuild-example/package.json
@@ -5,8 +5,8 @@
   "description": "Simple esbuild example for @stylexjs/esbuild-plugin",
   "main": "src/App.jsx",
   "scripts": {
-    "build": "node scripts/build.mjs",
-    "lint": "eslint \"**/*.{js,jsx}\""
+    "app:build": "node scripts/build.mjs",
+    "app:lint": "eslint \"**/*.{js,jsx}\""
   },
   "license": "MIT",
   "dependencies": {
