+++ package.json
@@ -4,17 +4,16 @@
   "version": "0.0.0",
   "license": "MIT",
   "scripts": {
+    "postinstall": "npm run build",
+    "prepare": "husky install tools/husky",
     "build": "npm run build --workspaces --if-present",
     "flow": "flow",
     "prettier": "prettier --write \"**/*.js\" \"**/*.ts\"",
     "prettier:report": "prettier --check \"**/*.js\"",
     "lint": "npm run lint:report -- --fix",
-    "lint:report": "eslint packages apps/docs apps/rollup-example apps/webpack-example",
-    "postinstall": "npm run build",
-    "precommit": "npm run prettier && npm run flow && npm run lint && npm run test",
-    "pretest": "npm run build",
-    "test": "npm run test --workspaces --if-present",
-    "test:nobuild": "npm run test --workspaces --if-present"
+    "lint:report": "eslint packages apps/docs apps/esbuild-example apps/rollup-example apps/webpack-example",
+    "test:packages": "npm run test --workspaces --if-present",
+    "test": "npm run build && npm run flow && npm run test:packages && npm run prettier:report && npm run lint:report"
   },
   "workspaces": [
     "packages/shared",
@@ -51,11 +50,14 @@
     "eslint-plugin-react": "^7.37.1",
     "flow-bin": "^0.261.0",
     "hermes-eslint": "^0.26.0",
+    "husky": "^8.0.0",
     "jest": "^30.0.0-alpha.6",
+    "lint-staged": "^13.0.3",
     "prettier": "^3.5.0",
     "prettier-plugin-hermes-parser": "^0.26.0",
     "rimraf": "^5.0.10",
     "rollup": "^4.24.0",
+    "serve": "^14.2.4",
     "typescript": "^5.3.3"
   },
   "prettier": {
@@ -78,7 +80,11 @@
       }
     ]
   },
-  "dependencies": {
-    "serve": "^14.2.4"
+  "lint-staged": {
+    "**/*.js": [
+      "prettier --write",
+      "git update-index --again",
+      "eslint"
+    ]
   }
 }
