+++ packages/benchmarks/package.json
@@ -1,0 +1,21 @@
+{
+  "private": true,
+  "name": "benchmarks",
+  "version": "0.10.1",
+  "scripts": {
+    "size": "NODE_ENV=production rollup --config ./rollup-size.config.mjs && node size/run.js",
+    "size:compare": "node size/compare.js"
+  },
+  "license": "MIT",
+  "dependencies": {
+    "@stylexjs/rollup-plugin": "0.10.1",
+    "clean-css": "^5.3.3",
+    "brotli-size": "^4.0.0",
+    "rollup": "^4.22.4",
+    "terser": "^5.3.0",
+    "yargs": "17.7.2"
+  },
+  "engines": {
+    "node": ">=20.11.0"
+  }
+}
