--- typos.toml
@@ -1,11 +1,0 @@
-[files]
-extend-exclude = [
-  "*.snap",
-  "apps/rollup-example/lotsOfStyles.js", 
-  "apps/rollup-example/lotsOfStylesDynamic.js", 
-  "flow-typed/*",
-]
-
-[default.extend-words]
-# Ignore false-positives
-stylex = "stylex"
